pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "insightface>=0.7.3",
    "numpy>=2.3.2",
    "onnxruntime>=1.22.1",
    "opencv-python>=4.11.0.86",
    "pillow==10.2.0",
    "python-multipart==0.0.9",
    "uvicorn>=0.35.0",
]
